% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramer.R
\docType{package}
\name{cramer-package}
\alias{cramer}
\alias{cramer-package}
\title{Cramer-Test for uni- and multivariate two-sample-problem}
\description{
Provides R routine for the so called two-sample Cramer-Test. This nonparametric two-sample-test on equality of the underlying distributions can be applied to multivariate data as well as univariate data. It offers two possibilities to approximate the critical value both of which are included in this package.
}
\author{
Carsten Franz <carsten.franz@gmail.com>
}
