\name{print.cramertest}
\alias{print.cramertest}
\title{
Print a Summary of a cramertest Object
}
\description{
This is a method for the function print for objects of the class \code{"cramertest"}.
}
\usage{
print.cramertest(cramer.test.out)
}
\arguments{
\item{cramer.test.out}{
A Cram\eqn{\mbox{\'e}}{e}r-test output object of class \code{"cramertest"} generated by the \code{cramer.test} function.
}}
\value{
The \code{cramertest} object is returned invisibly.
}
\details{
Prints a test summary containing information about the dimension of the samples, their
sizes, the resulting test statistic as well as the critical value, the number of 
replicates and the bootstrap-method used to obtain the critical value. Finally the 
result of the test, whether the hypothesis of same distributions should be accepted or not, is 
shown.
}
\seealso{
\code{\link{cramer.test}}
}
\keyword{nonparametric}
\keyword{cramer}
\keyword{two-sample}
\keyword{twosample}
	                                                   